cms.food|member|upload and api.v1.food|member done 1.0.0

+++ app/models/stat_daily_site.py
@@ -1,14 +1,13 @@
-from sqlalchemy import Column, Integer, Float
+from sqlalchemy import Column, Integer, DECIMAL
 
-from app.models.base import Base
+from . import Base
 
 
 class StatDailySite(Base):
     id = Column(Integer, primary_key=True)
     date = Column(Integer, nullable=False, comment='统计日期')
-    total_money = Column(Float, nullable=False, comment='当日应收总金额')
+    total_money = Column(DECIMAL, nullable=False, comment='当日应收总金额')
     total_member_count = Column(Integer, nullable=False, comment='当日会员总数')
     total_new_member_count = Column(Integer, nullable=False, comment='当日新增会员总数')
     total_order_count = Column(Integer, nullable=False, comment='当日订单总数')
     total_share_count = Column(Integer, nullable=False, comment='当日分享总数')
-
